Refactor PermissionsManager
